Update README.md
I felt the need to understand better the Gitcoin implementation.  ht to Alek who pointed me in this direction and via Discord provided some explanation - you guessed it - ChatGPT.

Hopefully, this readme could be useful to others in the future. Or it could be totally in the wrong place etc - feel free to reject it or whatever.